fix(examples): update manual cost-aware BO example and remove test_ prefix
