+++ examples/6_cost_aware/1_cost_aware_facade.py
