+++ examples/6_cost_aware/2_initial_design.py
