+++ examples/6_cost_aware/3_ei_cool_acquisition.py
